Halloween Trick or Treating Analysis with Story
My analysis of the Halloween Trick or Treating dataset that we used for class. The dataset gave information about Trick or Treating trends from 2008 - 2019, including the days of the week that Halloween occurred and the times that Trick or Treaters showed up. I show the most popular Trick or Treating trends by day and time in this Tableau file.