Analyzing International Debt Statistics
This is a guided SQL project from DataCamp that analyzes a dataset of international debt information from The World Bank for 124 different countries. The project determines which country in the data set has the most debt and for which debt category, and also explores the amounts and types of debt in the dataset.